Move ActivationFacory into WinRT.Runtime.dll
