+++ src/WinRT.Runtime/Context.cs
@@ -10,6 +10,16 @@ namespace WinRT
     static partial class Context
     {
         [DllImport("api-ms-win-core-com-l1-1-0.dll")]
+        private static extern unsafe int CoGetContextToken(IntPtr* contextToken);
+
+        public unsafe static IntPtr GetContextToken()
+        {
+            IntPtr contextToken;
+            Marshal.ThrowExceptionForHR(CoGetContextToken(&contextToken));
+            return contextToken;
+        }
+
+        [DllImport("api-ms-win-core-com-l1-1-0.dll")]
         private static extern unsafe int CoGetObjectContext(Guid* riid, IntPtr* ppv);
 
         private static readonly Guid IID_ICallbackWithNoReentrancyToApplicationSTA = new(0x0A299774, 0x3E4E, 0xFC42, 0x1D, 0x9D, 0x72, 0xCE, 0xE1, 0x05, 0xCA, 0x57);
